+++ components/atom/button/src/index.js
@@ -5,7 +5,23 @@ import cx from 'classnames'
 import Button from './Button'
 
 const CLASS = 'sui-AtomButton'
+
 const TYPES = ['primary', 'accent', 'secondary', 'tertiary']
+const DESIGNS = ['solid', 'outline', 'flat']
+const COLORS = [
+  'primary',
+  'accent',
+  'neutral',
+  'success',
+  'alert',
+  'error',
+  'social-facebook',
+  'social-twitter',
+  'social-google',
+  'social-youtube',
+  'social-whatsapp',
+  'social-instagram'
+]
 const GROUP_POSITIONS = {
   FIRST: 'first',
   MIDDLE: 'middle',
@@ -16,20 +32,34 @@ const MODIFIERS = ['disabled', 'fullWidth', 'focused', 'negative', 'link']
 const OWN_PROPS = [
   ...TYPES,
   ...SIZES,
+  'children',
+  'className',
+  'color',
+  'design',
+  'focused',
+  'fullWidth',
   'groupPosition',
   'leftIcon',
-  'rightIcon',
-  'className',
-  'children',
-  'fullWidth',
-  'focused',
   'negative',
+  'rightIcon',
   'type'
 ]
-const CLASSES = [...TYPES, ...SIZES, ...MODIFIERS, 'empty'].reduce(
-  (res, key) => Object.assign(res, {[key]: `${CLASS}--${key}`}),
-  {}
-)
+
+const createClasses = (array, sufix = '') => {
+  return array.reduce(
+    (res, key) => ({...res, [key]: `${CLASS}--${key}${sufix}`}),
+    {}
+  )
+}
+
+const CLASSES = createClasses([
+  ...TYPES,
+  ...DESIGNS,
+  ...SIZES,
+  ...MODIFIERS,
+  'empty'
+])
+const COLOR_CLASSES = createClasses(COLORS, 'Color')
 
 /**
  * Get props cleaning out AtomButton own props
@@ -53,27 +83,50 @@ const getModifiers = props => {
   )
 }
 
+const getPropsWithDefaultValues = props => {
+  let {color, design, type} = props
+  // if color or design are defined, use them with the passed or default value
+  if (color || design) {
+    color = color || 'primary'
+    design = design || 'solid'
+  } else {
+    type = type || 'primary'
+  }
+
+  return {
+    color,
+    design,
+    type,
+    ...props
+  }
+}
+
 const AtomButton = props => {
   const {
+    color,
     children,
     className,
     groupPosition,
     leftIcon,
     rightIcon,
     size,
+    design,
     title,
     type
-  } = props
+  } = getPropsWithDefaultValues(props)
 
   const classNames = cx(
     CLASS,
-    CLASSES[type],
+    !type && COLOR_CLASSES[color],
+    !type && CLASSES[design],
+    type && CLASSES[type],
     groupPosition && `${CLASS}-group ${CLASS}-group--${groupPosition}`,
     size && CLASSES[size],
     getModifiers(props).map(key => CLASSES[key]),
     !children && CLASSES.empty,
     className
   )
+
   const newProps = cleanProps(props)
 
   return (
@@ -95,6 +148,22 @@ AtomButton.displayName = 'AtomButton'
 
 AtomButton.propTypes = {
   /**
+   * Color of button:
+   * 'primary' (default),
+   * 'accent',
+   * 'neutral',
+   * 'success',
+   * 'alert',
+   * 'error',
+   * 'social-facebook',
+   * 'social-twitter',
+   * 'social-google',
+   * 'social-youtube',
+   * 'social-whatsapp',
+   * 'social-instagram'
+   */
+  color: PropTypes.oneOf(COLORS),
+  /**
    * HTML element: if true, render a link. Otherwise render a button
    */
   link: PropTypes.bool,
@@ -111,9 +180,19 @@ AtomButton.propTypes = {
    */
   title: PropTypes.string,
   /**
-   * Type of button: 'primary' (default), 'accent', 'secondary', 'tertiary'
+   * DEPRECATED. Type of button: 'primary' (default), 'accent', 'secondary', 'tertiary'
    */
-  type: PropTypes.oneOf(TYPES),
+  type: function(props, propName, componentName) {
+    if (props[propName] !== undefined) {
+      console.warn(
+        `The prop ${propName} is DEPRECATED on ${componentName}. You should use now "design" and "color" props.`
+      )
+    }
+  },
+  /**
+   * Design style of button: 'solid' (default), 'outline', 'flat'
+   */
+  design: PropTypes.oneOf(DESIGNS),
   /**
    * Group position: 'first', 'middle' (default), 'last'
    */
@@ -171,10 +250,7 @@ AtomButton.propTypes = {
   isButton: PropTypes.bool
 }
 
-AtomButton.defaultProps = {
-  type: 'primary'
-}
-
 export default AtomButton
 export {GROUP_POSITIONS as atomButtonGroupPositions}
 export {TYPES as atomButtonTypes}
+export {COLORS as atomButtonColors}